applications/machine_learning/object_classification/src/database/pyproject.toml
[project]
dependencies = ["aio-pika>=9.5.5", "motor>=3.7.1", "pyyaml>=6.0.2"]

name = "ml_consumer"
version = "0.1.0"
requires-python = ">= 3.10.12"

[tool.rye]
managed = true
dev-dependencies = []
virtual = true

[tool.hatch.metadata]
allow-direct-references = true
